+++ src/pebble/domain/entities/habit_category.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Optional
 
-from pebble.domain.value_objects import Color
-from pebble.domain.value_objects.types import ID, Description, Name
+from ..value_objects.color import Color
+from ..value_objects.types import ID, Description, Name
 
 
 @dataclass